Change the location of the router interfaces
