--- src/Framework/Interfaces/Router/Exception/NotFoundException.php
@@ -1,24 +1,0 @@
-<?php
-/**
- * PHP Version 5.6.0
- *
- * @category Errors
- * @package  Onion\Framework\Interfaces\Router\Exception
- * @author   Dimitar Dimitrov <daghostman.dd@gmail.com>
- * @license  https://opensource.org/licenses/MIT MIT License
- * @link     https://github.com/phOnion/framework
- */
-
-namespace Onion\Framework\Interfaces\Router\Exception;
-
-/**
- * Interface NotFoundException.
- * Exception to indicate that a `404 not found`-type error occurred, while
- * performing the routing.
- *
- * @package Onion\Framework\Interfaces\Router\Exception
- */
-interface NotFoundException
-{
-
-}
