+++ src/Framework/Router/Interfaces/Exception/NotFoundException.php
@@ -1,0 +1,15 @@
+<?php
+declare(strict_types = 1);
+namespace Onion\Framework\Router\Interfaces\Exception;
+
+/**
+ * Interface NotFoundException.
+ * Exception to indicate that a `404 not found`-type error occurred, while
+ * performing the routing.
+ *
+ * @package Onion\Framework\Interfaces\Router\Exception
+ */
+interface NotFoundException
+{
+
+}
